requirements.txt
# local package
-e .

# external requirements
simpy~=4.0.1
numpy~=1.21.2
pandas~=1.3.5
networkx~=2.6.3
statsmodels~=0.13.0
pyyaml~=6.0
scipy~=1.7.3
sphinx
pytest
